Allow server restarts without rebooting client
Now, if the labview server is stopped and restarted (to change the
binning settings, for example), the ccd_client program won't need to be
restarted. Just press pause on the client before stopping the server,
and hit 'resume' once the server is back up.

Requires wanglib at f13b0ab007b9 or better.