Fixed problem with checkbox not working
Root cause: top-level Frame class was adding panels to sub-panels'
sizers. That's sloppy - and it doesn't work on windows, apparently.

Maybe I could have gotten away with re-calling 'fit' on the subsizers,
but it's better to just do it right. I created a sidebar sizer that can
house the top-level stuff.